moved classes into the Saucedemo package
