+++ src/main/java/PageObjects/Saucedemo/CartPage.java
@@ -1,9 +1,7 @@
-package PageObjects;
+package PageObjects.Saucedemo;
 
 import org.openqa.selenium.By;
 import org.testng.Assert;
-
-import java.util.Locale;
 
 public class CartPage extends BasePage {
