+++ src/main/java/PageObjects/Saucedemo/CheckoutOverviewPage.java
@@ -1,4 +1,4 @@
-package PageObjects;
+package PageObjects.Saucedemo;
 
 import org.openqa.selenium.By;
