+++ src/main/java/PageObjects/Saucedemo/CheckoutPage.java
@@ -1,8 +1,6 @@
-package PageObjects;
+package PageObjects.Saucedemo;
 
 import org.openqa.selenium.By;
-import org.testng.Assert;
-import java.util.Locale;
 
 public class CheckoutPage extends BasePage {
