+++ src/main/java/PageObjects/Saucedemo/ItemPage.java
@@ -1,4 +1,4 @@
-package PageObjects;
+package PageObjects.Saucedemo;
 
 import org.openqa.selenium.By;
