+++ src/test/java/Task_9.java
@@ -1,4 +1,4 @@
-import PageObjects.*;
+import PageObjects.Saucedemo.*;
 import driver.BaseTest;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
